CS61A
The CS 61 series is an introduction to computer science, with particular emphasis on software and on machines from a programmer's point of view.  

CS 61A concentrates on the idea of abstraction, allowing the programmer to think in terms appropriate to the problem rather than in low-level operations dictated by the computer hardware.
CS 61B deals with the more advanced engineering aspects of software, such as constructing and analyzing large programs.
CS 61C focuses on machines and how they execute the programs you write.
In CS 61A, we are interested in teaching you about programming, not about how to use one particular programming language. We consider a series of techniques for controlling program complexity, such as functional programming, data abstraction, and object-oriented programming.

CS 61A primarily uses the Python 3 programming language. Python is a popular language in both industry and academia. It is also particularly well-suited to the task of exploring the topics taught in this course. It is an open-source language developed by a large volunteer community that prides itself on the diversity of its contributors. We will also examine the Scheme programming language, as well as the Structured Query Language (SQL) in the latter half of the course.

Mastery of a particular programming language is a very useful side effect of CS 61A. However, our goal is not to choose what language you use in your future studies and career. Instead, our hope is that once you have learned the essence of programming, you will find that picking up a new programming language is but a few days' work.